Moved firebase initialization methods into new class
