+++ app/src/main/java/com/example/recalltracker/MainActivity.java
@@ -1,82 +1,20 @@
 package com.example.recalltracker;
 
-import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
-import android.util.Log;
-import android.widget.Toast;
 
-import com.example.recalltracker.Utils.DatabaseAPI;
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.Task;
-import com.google.firebase.auth.AuthResult;
-import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.firestore.CollectionReference;
-import com.google.firebase.firestore.FirebaseFirestore;
-import com.google.firebase.iid.FirebaseInstanceId;
-import com.google.firebase.iid.InstanceIdResult;
+import com.example.recalltracker.Utils.Firebase;
 
-import java.util.ArrayList;
-import java.util.Map;
-import java.util.Objects;
 
 public class MainActivity extends AppCompatActivity {
-
-    private static final String TAG = "MainActivity";
-    private FirebaseAuth mAuth;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        mAuth = FirebaseAuth.getInstance();
-        FirebaseInstanceId.getInstance().getInstanceId()
-                .addOnCompleteListener(new OnCompleteListener<InstanceIdResult>() {
-                    @Override
-                    public void onComplete(@NonNull Task<InstanceIdResult> task) {
-                        if (!task.isSuccessful()) {
-                            Log.w(TAG, "getInstanceId failed", task.getException());
-                            return;
-                        }
-
-                        // Get new Instance ID token
-                        String token = Objects.requireNonNull(task.getResult()).getToken();
-                        Log.d(TAG, "Token: " + token);
-
-                        // Log and toast
-                        Log.d(TAG, "Retrieved firebase push token");
-                        Toast.makeText(MainActivity.this,
-                                "Retrieved firebase push token", Toast.LENGTH_SHORT).show();
-
-                        firebaseSignIn(token);
-                    }
-                });
-
-    }
-
-    public void firebaseSignIn(final String pushToken) {
-        super.onStart();
-        mAuth.signInAnonymously()
-                .addOnCompleteListener(this, new OnCompleteListener<AuthResult>() {
-                    @Override
-                    public void onComplete(@NonNull Task<AuthResult> task) {
-                        if (task.isSuccessful()) {
-                            String userId = Objects.requireNonNull(mAuth.getCurrentUser()).getUid();
-                            FirebaseFirestore db = FirebaseFirestore.getInstance();
-                            CollectionReference users = db.collection("users");
-
-                            Map<String, Object> data = DatabaseAPI.createUserData(pushToken,
-                                    true, new ArrayList<String>());
-
-                            DatabaseAPI.updateUser(users, userId, data);
-                        } else {
-                            Toast.makeText(MainActivity.this, "Authentication failed.",
-                                    Toast.LENGTH_SHORT).show();
-                        }
-                    }
-                });
+        Firebase.firebaseInit(this);
     }
 
 }
